--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -111,8 +111,12 @@ jobs:
           - "1"    # Latest release
         os:
           - ubuntu-latest
+        tzjdata-version:
+          - "1.3.0"  # Version which does not support `artifact_dir`
+          - "1"
     env:
       JULIA_DEPOT_PATH: build-depot
+      TZJDATA_VERSION: ${{ matrix.tzjdata-version }}
     steps:
       - uses: actions/checkout@v4
       - uses: julia-actions/setup-julia@v2
@@ -123,6 +127,7 @@ jobs:
         shell: julia --color=yes --project=sysimage {0}
         run: |
           using Pkg
+          Pkg.add(PackageSpec(name="TZJData", version=ENV["TZJDATA_VERSION"]))
           Pkg.add(PackageSpec(name="TimeZones", rev=ENV["SHA"]))
       - name: Create sysimage
         shell: julia --color=yes {0}