--- a/Project.toml
+++ b/Project.toml
@@ -4,6 +4,7 @@ authors = ["Curtis Vogt <curtis.vogt@gmail.com>"]
 version = "1.18.1"
 
 [deps]
+Artifacts = "56f22d72-fd6d-98f1-02f0-08ddc0907c33"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 InlineStrings = "842dd82b-1e85-43dc-bf29-5d0ee9dffc48"
@@ -22,6 +23,7 @@ RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 TimeZonesRecipesBaseExt = "RecipesBase"
 
 [compat]
+Artifacts = "1"
 Aqua = "0.8"
 Dates = "1"
 Downloads = "1"