--- a/src/TimeZones.jl
+++ b/src/TimeZones.jl
@@ -43,10 +43,22 @@ abstract type Local <: TimeZone end
 function __init__()
     # Must be set at runtime to ensure relocatability 
     _COMPILED_DIR[] = if isdefined(TZJData, :artifact_dir)
+        # Recent versions of TZJData are relocatable
         TZJData.artifact_dir()
     else
-        # Backwards compatibility with TZJData v1.3.0 and below.
-        Artifacts.artifact_path(Base.SHA1(basename(TZJData.ARTIFACT_DIR)))
+        # Backwards compatibility with TZJData v1.3.0 and below
+        hash = if TZJData.TZDATA_VERSION == "2024b"
+            Base.SHA1("7fdea2a12522469ca39925546d1fd93c10748180")
+        elseif TZJData.TZDATA_VERSION == "2024a"
+            Base.SHA1("520ce3f83be7fbb002cca87993a1b1f71fe10912")
+        elseif TZJData.TZDATA_VERSION == "2023d"
+            Base.SHA1("b071cffdb310f5d3ca640c09cfa3dc3f23d450ad")
+        elseif TZJData.TZDATA_VERSION == "2023c"
+            Base.SHA1("52e48e96c4df04eeebc6ece0d9f1c3b545f0544c")
+        else
+            error("TZJData.jl with TZDATA_VERSION = $(TZJData.TZDATA_VERSION) is supposed to be relocatable!")
+        end
+        Artifacts.artifact_path(hash)
     end
 
     # Dates extension needs to happen everytime the module is loaded (issue #24)