--- a/src/TimeZones.jl
+++ b/src/TimeZones.jl
@@ -41,7 +41,6 @@ const _COMPILED_DIR = Ref{String}()
 abstract type Local <: TimeZone end
 
 function __init__()
-    # COV_EXCL_START
     # Set at runtime to ensure relocatability
     _COMPILED_DIR[] = @static if isdefined(TZJData, :artifact_dir)
         TZJData.artifact_dir()
@@ -51,7 +50,6 @@ function __init__()
         hash = Base.SHA1(artifact_dict["tzjdata"]["git-tree-sha1"])
         Artifacts.artifact_path(hash)
     end
-    # COV_EXCL_STOP
 
     # Dates extension needs to happen everytime the module is loaded (issue #24)
     init_dates_extension()