Fixes #467 

Half of a possible solution to https://github.com/JuliaTime/TimeZones.jl/issues/467. The other half is at https://github.com/JuliaTime/TZJData.jl/pull/32. 

As discussed in the issue, there are many possible ways to fix this. I have tested that this approach works when including `TimeZones.jl` in a system image and relocating the depot. I think it should also be 100% equivalent to the old code in a normal setup. I cannot exclude that there are other relocatability issues hidden away, but even if this will not be the long term solution to the problem I think it is a decent fix for most people's use-cases.