@omus 2024-10-30
Would be good to re-run and post the benchmarks from https://github.com/JuliaTime/TimeZones.jl/pull/457 around importing to see if this has an impact on load time.