@lcontento 2024-10-30
> Would be good to re-run and post the benchmarks from #457 around importing to see if this has an impact on load time.

Sure, will do.