@lcontento 2024-12-17
I tried to avoid the code coverage errors by excluding the new code, but the project level check still fails. The errors on Julia nightly were not there last time I looked at this, but they do not seem related to this PR.
@omus Is there anything else you would like to see improved here?