@omus 2024-12-20
I think I would prefer to take the CodeCov hit and try to address the coverage drop later