@omus 2024-12-20
I re-ran the benchmark on Julia 1.11.2:

With this PR:
```
julia> @time_imports import TimeZones
      0.7 ms  Printf
     17.2 ms  Dates
      0.4 ms  Scratch
      3.7 ms  InlineStrings
      0.5 ms  TZJData
               ┌ 0.1 ms p7zip_jll.__init__()
      3.7 ms  p7zip_jll
               ┌ 0.0 ms NetworkOptions.__init__()
      2.3 ms  NetworkOptions
      8.0 ms  ArgTools
               ┌ 0.3 ms nghttp2_jll.__init__()
      2.3 ms  nghttp2_jll
               ┌ 1.6 ms LibCURL_jll.__init__()
      3.4 ms  LibCURL_jll
               ┌ 0.0 ms MozillaCACerts_jll.__init__()
      2.0 ms  MozillaCACerts_jll
               ┌ 0.0 ms LibCURL.__init__()
      1.1 ms  LibCURL
               ┌ 2.9 ms Downloads.Curl.__init__()
     15.7 ms  Downloads
      0.3 ms  UUIDs
      0.4 ms  Compat
      0.3 ms  Compat → CompatLinearAlgebraExt
      0.4 ms  ExprTools
      0.5 ms  Mocking
               ┌ 3.6 ms TimeZones.TZData.__init__() 62.16% compilation time
               ├ 0.1 ms TimeZones.__init__()
     24.9 ms  TimeZones 8.91% compilation time

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  73.451 ns (1 allocation: 48 bytes)
```

Before this PR (with updated TZJData):
```
julia> @time_imports import TimeZones
      0.6 ms  Printf
     14.4 ms  Dates
      0.5 ms  Scratch
      2.9 ms  InlineStrings
      0.4 ms  TZJData
               ┌ 0.0 ms p7zip_jll.__init__()
      3.1 ms  p7zip_jll
               ┌ 0.0 ms NetworkOptions.__init__()
      1.8 ms  NetworkOptions
      7.2 ms  ArgTools
               ┌ 0.4 ms nghttp2_jll.__init__()
      2.4 ms  nghttp2_jll
               ┌ 1.6 ms LibCURL_jll.__init__()
      3.4 ms  LibCURL_jll
               ┌ 0.0 ms MozillaCACerts_jll.__init__()
      2.3 ms  MozillaCACerts_jll
               ┌ 0.0 ms LibCURL.__init__()
      1.4 ms  LibCURL
               ┌ 2.8 ms Downloads.Curl.__init__()
     16.3 ms  Downloads
      0.4 ms  UUIDs
      0.4 ms  Compat
      0.3 ms  Compat → CompatLinearAlgebraExt
      0.4 ms  ExprTools
      0.5 ms  Mocking
               ┌ 2.7 ms TimeZones.TZData.__init__() 79.32% compilation time
               ├ 0.0 ms TimeZones.__init__()
     24.2 ms  TimeZones 8.97% compilation time

julia> using BenchmarkTools, TimeZones

julia> @btime istimezone("Europe/Warsaw");
  73.484 ns (1 allocation: 48 bytes)
```