@omus 2024-12-20
Hmmm, the CI sysimage test isn't failing when I expected it to