@omus 2024-12-20
I've verified that the tests are failing when using `pkgdir` and older versions of TZJData: https://github.com/JuliaTime/TimeZones.jl/pull/479/commits/73da68afd6d40266b3d56a1df8f71c038a8ff9c5. Should be good to merge this after the CI wraps up. Thanks everyone!